Make testing utils importable (#1037)
* fixed pre-commit warning

* moved testing utils for vasp

* simplied tutorial mock vasp import

* docs and name cleanup

---------
